permissions

+++ app/Http/Controllers/EquipesController.php
@@ -12,20 +12,38 @@ use App\Models\Access\User\User as user;
 class EquipesController extends Controller
 {
     public function index(){
+      if(!isAllowed('ver-equipes') && !isAllowed('ver-equipes-self')){
+         \Session::flash('flash_danger', 'Você não tem permissão para fazer isto.');
+         return redirect()->back();
+      } 
    		return view('backend.equipes');
    }
 
     public function getEquipes(){
     	$equipes = equipe::all();
     	$data = array();
-   		foreach($equipes as $equipe){
-   			$data[] = array(
-   				'nome'		 => '<a href="#" class="equipe-info" id="EID'.$equipe->id.'">'.$equipe->descricao.'</a>',
-   				'obs'   => !empty($equipe->obs) ? $equipe->obs : '-',
-   				'resp'   => !empty($equipe->responsavel_id) ? $equipe->responsavel->name : '-',
-   				'memb'  => $equipe->users->count()
-			);
-   		}
+      if(access()->user()->hasRole(1)){
+        foreach($equipes as $equipe){
+          $data[] = array(
+            'nome'     => '<a href="#" class="equipe-info" id="EID'.$equipe->id.'">'.$equipe->descricao.'</a>',
+            'obs'   => !empty($equipe->obs) ? $equipe->obs : '-',
+            'resp'   => !empty($equipe->responsavel_id) ? $equipe->responsavel->name : '-',
+            'memb'  => $equipe->users->count()
+        );
+        }
+      }else{
+        foreach($equipes as $equipe){
+          if($equipe->responsavel->id === access()->user()->id){
+            $data[] = array(
+              'nome'     => '<a href="#" class="equipe-info" id="EID'.$equipe->id.'">'.$equipe->descricao.'</a>',
+              'obs'   => !empty($equipe->obs) ? $equipe->obs : '-',
+              'resp'   => !empty($equipe->responsavel_id) ? $equipe->responsavel->name : '-',
+              'memb'  => $equipe->users->count()
+            );
+          }
+        }
+      }
+
    		$total = empty($equipes) ? 0 : $equipes->count();
    		$response['data'] = $data;
    		$response['current'] = 1;
@@ -36,6 +54,9 @@ class EquipesController extends Controller
     }
 
     public function info(Request $request){
+      if(!isAllowed('ver-equipes') && !isAllowed('ver-equipes-self')){
+        return view('backend.modals.unauthorized');
+      } 
   	$id = str_replace('EID', '', $request['id']);
   	$equipe = equipe::find($id);
     $membros = array();
@@ -46,6 +67,9 @@ class EquipesController extends Controller
   }
 
   public function infoSmall(Request $request){
+    if(!isAllowed('ver-equipes') && !isAllowed('ver-equipes-self')){
+      return view('backend.modals.unauthorized');
+    }
     $id = str_replace('EID', '', $request['id']);
     $equipe = equipe::find($id);
 
@@ -53,16 +77,25 @@ class EquipesController extends Controller
   }
 
   public function criar(){
+    if(!isAllowed('criar-equipes')){
+      return view('backend.modals.unauthorized');
+    }
       return view('backend.modals.equipes.criar');
    }
 
    public function editar(Request $request){
+    if(!isAllowed('criar-equipes')){
+      return view('backend.modals.unauthorized');
+    }
     $id = str_replace('EDI', '', $request['id']);
     $equipe = equipe::find($id);
     return view('backend.modals.equipes.edit', compact('equipe'));
   }
 
   public function update(Request $request){
+    if(!isAllowed('criar-equipes')){
+      return '%error&Você não tem permissão para fazer isto.';
+    }
     $Alldados = $request->all();
     $dadosBefore = urldecode($Alldados['dados']);
     $dados = explode('&', $dadosBefore);
@@ -75,6 +108,8 @@ class EquipesController extends Controller
     $id = $check['id'];
     unset($check['id']);
     $new = equipe::find($id);
+    $userRole = user::find($check['responsavel_id']); 
+    $userRole->handleLider();
     if($new->responsavel_id != $check['responsavel_id']){
       // $oldResp = user::find($new->responsavel_id);
       // $new->users()->detach($oldResp);
@@ -89,14 +124,22 @@ class EquipesController extends Controller
   }
 
   public function delete(Request $request){
+    if(!isAllowed('deletar-equipes')){
+      return '%error&Você não tem permissão para fazer isto.';
+    }
     $id = str_replace('DEI', '', $request['id']);
     $equipe = equipe::find($id);
+    $userRole = $equipe->responsavel;
+    $userRole->handleLider();
     $equipe->users()->detach();
     $equipe->delete();
     return '%success&Equipe Excluida com Sucesso';
   }
 
   public function novoMembro(Request $request){
+    if(!isAllowed('criar-equipes') && !isAllowed('add-equipe-self')){
+      return '%error&Você não tem permissão para fazer isto.';
+    }
     $id = $request['id'];
     $eid = $request['equipe_id'];
     $equipe = equipe::find($eid);
@@ -114,6 +157,9 @@ class EquipesController extends Controller
   }
 
   public function removerMembro(Request $request){
+    if(!isAllowed('criar-equipes') && !isAllowed('add-equipe-self')){
+      return '%error&Você não tem permissão para fazer isto.';
+    }
     $dados = $request->all();
     $id = $dados['id'];
     $eid = $dados['equipe_id'];
